Unify IndentedStringBuilder across MSTest source generators (#8987)
